Structure: Re-arrange hook folder structure
