+++ src/hooks/useAuth/index.ts
@@ -1,5 +1,1 @@
-// eslint-disable-next-line no-restricted-imports
-import { AuthContext, AuthContextType } from '@moneyboy/contexts/authContext';
-import { useContext } from 'react';
-
-export const useAuth: () => AuthContextType = () => useContext(AuthContext);
+export * from './useAuth.hook';
