+++ src/hooks/useAuth/useAuth.hook.ts
@@ -1,0 +1,5 @@
+// eslint-disable-next-line no-restricted-imports
+import { AuthContext, AuthContextType } from '@moneyboy/contexts/authContext';
+import { useContext } from 'react';
+
+export const useAuth: () => AuthContextType = () => useContext(AuthContext);
