+++ src/hooks/useSettings/useSettings.hook.ts
@@ -1,0 +1,5 @@
+// eslint-disable-next-line no-restricted-imports
+import { SettingsContext, SettingsContextType } from '@moneyboy/contexts/settingsContext';
+import { useContext } from 'react';
+
+export const useSettings: () => SettingsContextType = () => useContext(SettingsContext);
