+++ src/hooks/useStyle/index.ts
@@ -1,5 +1,1 @@
-// eslint-disable-next-line no-restricted-imports
-import { StyleContext, StyleContextType } from '@moneyboy/contexts/styleContext';
-import { useContext } from 'react';
-
-export const useStyle: () => StyleContextType = () => useContext(StyleContext);
+export * from './useStyle.hook';
