+++ src/hooks/useStyle/useStyle.hook.ts
@@ -1,0 +1,5 @@
+// eslint-disable-next-line no-restricted-imports
+import { StyleContext, StyleContextType } from '@moneyboy/contexts/styleContext';
+import { useContext } from 'react';
+
+export const useStyle: () => StyleContextType = () => useContext(StyleContext);
